+++ packages/gensx/tests/commands/environment/unselect.test.ts
@@ -1,5 +1,3 @@
-import fs from "node:fs/promises";
-import os from "node:os";
 import path from "node:path";
 
 import { render } from "ink-testing-library";
@@ -19,11 +17,18 @@ import { UnselectEnvironmentUI } from "../../../src/commands/environment/unselec
 import * as projectModel from "../../../src/models/projects.js";
 import * as envConfig from "../../../src/utils/env-config.js";
 import * as projectConfig from "../../../src/utils/project-config.js";
+import {
+  cleanupProjectFiles,
+  cleanupTestEnvironment,
+  setupTestEnvironment,
+  waitForMockCall,
+  waitForText,
+} from "../../test-helpers.js";
 
-// Create a temporary directory for our tests
+// Setup test variables
 let tempDir: string;
-let origConfigDir: string | undefined;
 let origCwd: typeof process.cwd;
+let origConfigDir: string | undefined;
 
 // Mock dependencies
 vi.mock("../../../src/models/projects.js", () => ({
@@ -40,32 +45,14 @@ vi.mock("../../../src/utils/project-config.js", () => ({
 
 // Set up and tear down the test environment
 beforeAll(async () => {
-  // Save original process.cwd
-  origCwd = process.cwd;
-
-  // Create a temp directory for our tests
-  tempDir = await fs.mkdtemp(path.join(os.tmpdir(), "gensx-unselect-test-"));
-
-  // Create project and config directories
-  await fs.mkdir(path.join(tempDir, "project"), { recursive: true });
-  await fs.mkdir(path.join(tempDir, ".gensx", "projects"), { recursive: true });
-
-  // Override the config directory by mocking with environment variable
-  origConfigDir = process.env.GENSX_CONFIG_DIR;
-  process.env.GENSX_CONFIG_DIR = path.join(tempDir, ".gensx");
+  const setup = await setupTestEnvironment("unselect-test");
+  tempDir = setup.tempDir;
+  origCwd = setup.origCwd;
+  origConfigDir = setup.origConfigDir;
 });
 
 afterAll(async () => {
-  // Restore original environment
-  process.cwd = origCwd;
-  if (origConfigDir) {
-    process.env.GENSX_CONFIG_DIR = origConfigDir;
-  } else {
-    delete process.env.GENSX_CONFIG_DIR;
-  }
-
-  // Clean up temp directory
-  await fs.rm(tempDir, { recursive: true, force: true });
+  await cleanupTestEnvironment(tempDir, origCwd, origConfigDir);
 });
 
 beforeEach(() => {
@@ -75,69 +62,8 @@ beforeEach(() => {
 
 afterEach(async () => {
   vi.resetAllMocks();
-
-  // Clean up any test files after each test
-  try {
-    const projectsDir = path.join(tempDir, ".gensx", "projects");
-    const files = await fs.readdir(projectsDir);
-
-    for (const file of files) {
-      if (file !== ".gitkeep") {
-        await fs.unlink(path.join(projectsDir, file));
-      }
-    }
-
-    // Clean up project config file
-    try {
-      await fs.unlink(path.join(tempDir, "project", "gensx.yaml"));
-    } catch (_error) {
-      // Ignore if file doesn't exist
-    }
-  } catch (_error) {
-    // Ignore cleanup errors
-  }
+  await cleanupProjectFiles(tempDir);
 });
-
-function waitForText(
-  getFrame: () => string | undefined,
-  text: string | RegExp,
-  timeout = 1000,
-) {
-  return new Promise<void>((resolve, reject) => {
-    const start = Date.now();
-    function check() {
-      const frame = getFrame() ?? ""; // treat undefined as empty string
-      if (typeof text === "string" ? frame.includes(text) : text.test(frame)) {
-        resolve();
-      } else if (Date.now() - start > timeout) {
-        reject(new Error(`Timed out waiting for text: ${text}`));
-      } else {
-        setTimeout(check, 20);
-      }
-    }
-    check();
-  });
-}
-
-// Helper to wait for a mock function to be called
-function waitForMockCall(
-  mockFn: ReturnType<typeof vi.fn>,
-  timeout = 200,
-): Promise<void> {
-  return new Promise<void>((resolve, reject) => {
-    const start = Date.now();
-    function check() {
-      if (mockFn.mock.calls.length > 0) {
-        resolve();
-      } else if (Date.now() - start > timeout) {
-        reject(new Error(`Timed out waiting for mock function to be called`));
-      } else {
-        setTimeout(check, 20);
-      }
-    }
-    check();
-  });
-}
 
 suite("environment unselect command", () => {
   it("should unselect environment for an existing project", async () => {
